+++ docs/zh-cn/changelog.md
@@ -30,7 +30,7 @@
   - 🐞 修复了with沙箱在默认模式下用户自定义的强隔离变量`scopeProperties`在卸载时无法清空的问题。
   - 🐞 修复了with沙箱子应用的`window.onpopstate`兜底到主应用导致多次执行的问题。
   - 🐞 修复了craco子应用style元素插入顺序错误导致样式优先级混乱的问题，[issue 1071](https://github.com/micro-zoe/micro-app/issues/1071)。
-  - 🐞 修复了iframe沙箱下插件系统`escapeProperties`配置不生效的问题。
+  - 🐞 修复了在部分场景下子应用重写`Array.prototype.includes`导致死循环的问题。
 
 - **Update**
   - 🚀 优化了对于iframe沙箱`Node.parentNode`的处理方式。